refactor: Cleans up HaystackVariableSupport
This reworks the HaystackVariableSupport to align with the Loki plugin's use of CustomVariableSupport: https://github.com/grafana/grafana/blob/15293a2ceb083108c0f15490933db523aa915c7d/public/app/plugins/datasource/loki/LokiVariableSupport.ts#L10

Since I didn't really use any source example when implementing it initially, it ended up being kinda messy.
